apps/api/pyproject.toml
[project]
name = "omninew-shopify-ai-api"
version = "0.1.0"
description = "FastAPI backend for the OmniNew Shopify AI assistant MVP"
requires-python = ">=3.11"
dependencies = [
  "email-validator==2.2.0",
  "fastapi==0.116.1",
  "httpx==0.28.1",
  "pydantic==2.11.7",
  "pydantic-settings==2.10.1",
  "pyjwt==2.10.1",
  "uvicorn[standard]==0.35.0",
]

[dependency-groups]
dev = [
  "pytest==8.4.1",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.sdist]
include = [
  "app",
  "run.py",
  "tests",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
